dynamic estimated samples calc by view width

+++ src/org/mortalis/homeplayer/decoder/DecoderResult.java
@@ -2,6 +2,6 @@ package org.mortalis.homeplayer.decoder;
 
 public class DecoderResult {
   
-  public float[] samples;
+  public short[] samples;
   
 }
